\name{scaleboot-package}
\alias{scaleboot-package}
\docType{package}
\title{
Approximately Unbiased P-values via Multiscale Bootstrap
}
\description{
Calculating approximately unbiased (AU) p-values from
multiscale bootstrap probabilities.
}
\details{
For a complete
list, use \code{library(help="scaleboot")}.

The methodology is described in Shimodiara (2008). For the use of
\code{scaleboot}, Shimodaira (2008) may be referenced.


Further information is available in the following vignette:
\tabular{ll}{
\code{usesb} \tab Multiscale Bootstrap using Scaleboot Package\cr
}
}
\author{
Hidetoshi Shimodaira

Maintainer: Hidetoshi Shimodaira <shimo@i.kyoto-u.ac.jp>

I thank Paul A. Sheridan for his comments to improve the manual pages.
}
\references{
  Shimodaira, H. (2008). Testing Regions with Nonsmooth Boundaries via
  Multiscale Bootstrap, \emph{Journal of Statistical Planning and
    Inference}, 138, 1227-1241.
  (\url{http://dx.doi.org/10.1016/j.jspi.2007.04.001}).
}
\keyword{ package }
\keyword{ models }
\keyword{ nonparametric }

